Import changes from code review branches.

+++ source/blender/blenkernel/intern/constraint.c
@@ -5321,7 +5321,7 @@ static void transformcache_evaluate(bConstraint *con, bConstraintOb *cob, ListBa
   }
 
   /* Do not process data if using a render time procedural. */
-  if (BKE_cache_file_uses_render_procedural(cache_file, scene, (const int)DEG_get_mode(cob->depsgraph))) {
+  if (BKE_cache_file_uses_render_procedural(cache_file, scene, DEG_get_mode(cob->depsgraph))) {
     return;
   }
